VEM

Vortex Element Method library for computational fluid dynamics based on the evolution of vorticity

VEM.VortexBlobs

VEM.circulation_type

VEM.compute_field

VEM.compute_field!

VEM.dimension

VEM.radius_type

VEM.scalar_type

VEM.source_type

VEM.VortexBlobs — Type

VortexBlobs{Dimension, ScalarType<:AbstractFloat, CirculationType<:Union{ScalarType, SVector{Dimension, ScalarType}}}

Construct a distribution of 2D or 3D blobs.

The state of the distribution is completely specified by three properties:

The circulation carried by the blobs: circulation

The core radius of the blobs: radius

The coordinates of the blobs: source

Note that circulation is a scalar in 2D and a 3-vector in 3D.

Example

Construct a distribution of two 2D blobs:

using StaticArrays
using VEM

circulation = [1.23, 4.56]
radius = [0.45, 0.67]
source = [SVector(0.0, 0.0), SVector(0.1, 0.1)]

blobs = VortexBlobs(circulation, radius, source)

Construct a distribution of 3D blobs:

 using StaticArrays
 using VEM
 
 circulation = [SVector(1.23, 4.56, 7.89), SVector(3.21, 6.54, 9.87)]
 radius = [0.54, 0.67]
 source = [SVector(0.0, 0.0, 0.0), SVector(0.1, 0.1, 0.1)]

 blobs = VortexBlobs(circulation, radius, source)

VEM.circulation_type — Method

circulation_type(blobs::VortexBlobs)

Determine the type of the circulation carried by a blob.

Example

julia> circulation_type(blobs::VortexBlobs{2, Float32})
Float32

julia> circulation_type(blobs::VortexBlobs{3, Float16})
SVector{3, Float16}

VEM.compute_field! — Method

compute_field!(field, kernel, blobs, targets)

In-place version of compute_field.

The induced field is stored in field. The output type of kernel must be compatible with eltype(field).

VEM.compute_field — Method

compute_field(kernel, blobs, targets)

Compute a field defined by kernel induced by a distribution of blobs at targets.

The induced field is specified by a kernel function. The kernel computes the induction due to a single blob at a single target. The dimension of the output must be defined by extending kernel_output_size for the type of the kernel function. It should take kernel as input and return a StaticArrays.Size. Outputs of arbitrary tensor rank are supported in 2D and 3D.

Example

Compute the velocity due to three 2D blobs at four targets:

using LinearAlgebra
using StaticArrays
using VEM

function velocity(circulation, radius, source, target)
    r = target - source
    rSq = dot(r, r)

    f = cross(circulation, r)/(rSq*pi*2)
    g = 1 - exp(-rSq/(2*radius^2))

    return f*g
end

VEM.kernel_output_size(::typeof(velocity)) = Size(2)

circulations = [1.23, 4.56, 7.89]
radii = [0.45, 0.90, 1.80]
sources = [SVector(-3.0, 4.0), SVector(-5.0, 0.0), SVector(1.0, 2.0)] 
blobs = VortexBlobs(circulations, radii, sources)

targets = [SVector(3.0, 5.0), SVector(-3.0, 5.0), SVector(-3.0, -5.0), SVector(3.0, -5.0)]

velocity_field = compute_field(velocity, blobs, targets)

VEM.dimension — Method

dimension_type(blobs::VortexBlobs)

Determine the dimension of a blob.

Example

julia> dimension(blobs::VortexBlobs{2})
2

julia> dimension_type(blobs::VortexBlobs{3})
3

VEM.radius_type — Method

radius_type(blobs::VortexBlobs)

Determine the type of the radius of a blob.

Example

julia> radius_type(blobs::VortexBlobs{2, Float32})
Float32

julia> radius_type(blobs::VortexBlobs{3, Float16})
Float16

VEM.scalar_type — Function

scalar_type(blobs::VortexBlobs)

Determine the primitive type used for the data contained by the VortexBlobs struct.

Example

julia> scalar_type(blobs::VortexBlobs{2, Float32})
Float32

julia> scalar_type(blobs::VortexBlobs{3, Float16})
Float16

VEM.source_type — Method

source_type(blobs::VortexBlobs)

Determine the type of the source of a blob.

Example

julia> source_type(blobs::VortexBlobs{2, Float32})
SVector{2, Float32}

julia> source_type(blobs::VortexBlobs{3, Float16})
SVector{3, Float16}