API

VEM.AbstractInducedField

VEM.AbstractRedistributionKernel

VEM.AbstractVortexBlob

VEM.CartesianMesh

VEM.Float

VEM.FlowProperties

VEM.FlowProperties

VEM.GaussianVortexBlob

VEM.LagrangianFlowSolver

VEM.M4Prime

VEM.NumericalConfiguration

VEM.NumericalConfiguration

VEM.SimulationTime

VEM.SimulationTime

VEM.VelocityField

VEM.VorticityField

Base.zero

Base.zero

VEM.advance!

VEM.advection!

VEM.blob_center

VEM.blob_center!

VEM.blob_circulation

VEM.blob_circulation!

VEM.blob_dimension

VEM.blob_radius

VEM.blob_radius!

VEM.blob_scalar

VEM.cells_per_axis

VEM.current_time

VEM.diffusion!

VEM.direct_sum

VEM.direct_sum!

VEM.end_time

VEM.evolve!

VEM.field_scalar

VEM.induce

VEM.induce

VEM.interpolate_circulation

VEM.mesh_bounds

VEM.mesh_dimension

VEM.mesh_nodes

VEM.mesh_scalar

VEM.node_spacing

VEM.nodes_per_axis

VEM.population_control!

VEM.redistribution

VEM.redistribution_weight

VEM.running

VEM.support_radius

VEM.AbstractInducedField — Type

AbstractInducedField

An abstract type representing the type of field to be induced by a vortex blob.

VEM.AbstractRedistributionKernel — Type

AbstractRedistributionKernel

An abstract type representing a vortex blob redistribution kernel, which defines how weights are assigned based on distance normalized by the vortex blob core radius.

VEM.AbstractVortexBlob — Type

AbstractVortexBlob{Dimension, Scalar <: AbstractFloat}

Abstract type representing a vortex blob in Dimension-dimensional space, parameterized by a floating-point type Scalar.

VEM.CartesianMesh — Type

CartesianMesh{Dimension,Scalar<:AbstractFloat}(cells_per_axis::NTuple{Dimension,Int}, spacing::Scalar)

A Cartesian mesh for vortex blob methods.

Argumentss

`cellsperaxis: The number of cells along each axis.

spacing: The spacing between nodes in the mesh.

VEM.Float — Type

Float

Default floating-point type (Float32 or Float64) corresponding to the system's word size.

VEM.FlowProperties — Type

FlowProperties{Dimension,Float<:AbstractFloat}

A struct to hold the properties of the flow, such as density, kinematic viscosity, and freestream velocity. The Dimension type parameter indicates the dimensionality of the flow (e.g., 2D or 3D), while the Float type parameter specifies the floating-point type used for the properties (e.g., Float64).

Fields

density: The density of the fluid (must be positive).

kinematic_viscosity: The kinematic viscosity of the fluid (must be positive).

freestream_velocity`: The freestream velocity of the flow.

VEM.FlowProperties — Method

FlowProperties(density::Float=1.225, kinematic_viscosity::Float=1.46e-5, freestream_velocity::SA.SVector{Dimension,Float}=zero(SA.SVector{Dimension,Float})) where {Dimension,Float}

A constructor for FlowProperties that provides default values for air at sea level.

Arguments

density`: The density of the fluid (default is 1.225).

kinematic_viscosity: The kinematic viscosity of the fluid (default is 1.46e-5).

freestream_velocity: The freestream velocity of the flow (default is a zero vector).

Returns

An instance of FlowProperties with the specified properties.

VEM.GaussianVortexBlob — Type

GaussianVortexBlob(circulation::Union{Scalar,StaticArrays.SVector{Dimension,Scalar}}, center::StaticArrays.SVector{Dimension, Scalar}, radius::Scalar) where {Dimension, Scalar}
GaussianVortexBlob(circulation, center, radius)

A smooth vortex blob with a Gaussian vorticity distribution. It is characterized by an amount of circulation it carries, a position and a finite core radius.

Arguments

circulation: The circulation strength of the vortex blob.

center: The position of the vortex blob.

radius: The core radius of the vortex blob.

Returns

A GaussianVortexBlob instance.

VEM.LagrangianFlowSolver — Type

LagrangianFlowSolver{FP<:FlowProperties,ST<:SimulationTime,NC<:NumericalConfiguration}

A struct to hold the Lagrangian flow solver.

Fields

vortex_blobs: A vector of vortex blobs representing the flow field.

flow_properties: The properties of the flow being simulated.

simulation_time: The time configuration for the simulation.

numerical_configuration: The numerical configuration for the solver.

VEM.M4Prime — Type

M4Prime

M4' is a vortex blob redistribution kernel. It is defined by a piecewise function that assigns weights based on the distance from the center of the vortex blob, normalized by the blob's core radius. The M4' kernel has a support radius of 2, meaning that it assigns non-zero weights to points at a distance greater than 2 core radii.

VEM.NumericalConfiguration — Type

NumericalConfiguration{AdvectionScheme<:ODE.OrdinaryDiffEqCore.OrdinaryDiffEqAlgorithm, DiffusionScheme<:ODE.OrdinaryDiffEqCore.OrdinaryDiffEqAlgorithm, RedistributionKernel<:AbstractRedistributionKernel, Float<:AbstractFloat}

A struct to hold the numerical configuration for the solver, including the advection and diffusion time schemes, the redistribution kernel, mesh spacing, overlap ratio, and population control threshold.

Fields

advection_time_scheme: The time-stepping scheme for advection.

diffusion_time_scheme: The time-stepping scheme for diffusion.

redistribution_kernel: The kernel used for redistribution.

mesh_spacing: The spacing of the mesh.

overlap_ratio: The ratio of overlap between mesh cells.

population_control_threshold: The threshold for population control.

VEM.NumericalConfiguration — Method

NumericalConfiguration(mesh_spacing::Float; advection_time_scheme::ODE.OrdinaryDiffEqCore.OrdinaryDiffEqAlgorithm, diffusion_time_scheme::ODE.OrdinaryDiffEqCore.OrdinaryDiffEqAlgorithm, redistribution_kernel::AbstractRedistributionKernel, overlap_ratio::Float, population_control_threshold::Float) where {Float<:AbstractFloat}

A constructor for NumericalConfiguration that takes in the mesh spacing and optional parameters for the advection and diffusion time schemes, redistribution kernel, overlap ratio, and population control threshold. It returns an instance of NumericalConfiguration with the specified parameters.

Arguments

mesh_spacing: The spacing of the mesh (must be positive).

advection_time_scheme: The time-stepping scheme for advection (default is ODE.RK4()).

diffusion_time_scheme: The time-stepping scheme for diffusion (default is ODE.RK4()).

redistribution_kernel: The kernel used for redistribution (default is M4Prime()).

overlap_ratio: The ratio of overlap between mesh cells (default is 1.0).

population_control_threshold: The threshold for population control (default is 1e-6).

VEM.SimulationTime — Type

SimulationTime{Float<:AbstractFloat}(start_time::Float, time_step::Float, number_of_steps::Int)

A structure to manage the simulation time.

Fields

start_time: The initial time of the simulation (must be non-negative).

time_step: The time increment for each simulation step (must be positive).

number_of_steps: The total number of steps to simulate (must be positive).

current_step: The current step of the simulation (must be between 0 and number of steps).

VEM.SimulationTime — Method

SimulationTime(time_step::Float, number_of_steps::Int; start_time::Float=0.0, current_step::Int=0) where {Float<:AbstractFloat}

A constructor for SimulationTime that initializes the structure with the given parameters.

Arguments

time_step: The time increment for each simulation step.

number_of_steps: The total number of steps to simulate.

start_time: The initial time of the simulation (default is 0.0).

current_step: The current step of the simulation (default is 0).

Returns

An instance of SimulationTime initialized with the provided parameters.

VEM.VelocityField — Type

VelocityField <: AbstractInducedField

A singleton type representing a velocity field.

VEM.VorticityField — Type

VorticityField <: AbstractInducedField

A singleton type representing a vorticity field.

Base.zero — Method

Base.zero(::T) where {T<:AbstractVortexBlob}

Return a vortex blob with all data set to 0.

Arguments

::AbstractVortexBlob: An instance of a vortex blob.

Returns

A vortex blob instance with zero circulation, center, and radius.

Base.zero — Method

zero(::Type{GaussianVortexBlob{Dimension,Scalar}}) where {Dimension, Scalar}

Return a GaussianVortexBlob with all data set to 0.

Arguments

GaussianVortexBlob{Dimension,Scalar}: The type of a vortex blob..

Returns

A GaussianVortexBlob instance with zero circulation, center, and radius.

VEM.advance! — Method

advance!(time::SimulationTime)

Update the simulation time by incrementing the current_step by 1.

Arguments

time: An instance of SimulationTime to be updated.

VEM.advection! — Method

advection!(blobs, time, time_step; time_scheme=ODE.RK4())

Advance the position of a collection of blobs by a single time step according to the local velocity field induced by the collection. The positions of the blobs are updated in-place.

Arguments

blobs: An iterable collection of vortex blobs.

time: The time at the start of a time step.

time_step: The time increment over a single step.

time_scheme: (Optional) The time integration scheme to use. Defaults to OrdinaryDiffEq.RK4() (classical Runge-Kutta 4th order).

VEM.blob_center! — Method

blob_center!(blob::AbstractVortexBlob, new_center)

Update the position of the vortex blob.

Arguments

blob: The vortex blob.

new_center: The new vortex blob position.

VEM.blob_center — Method

blob_center(blob::AbstractVortexBlob)

Arguments

blob: The vortex blob.

Returns

The position of the vortex blob.

VEM.blob_circulation! — Method

blob_circulation!(blob::AbstractVortexBlob, new_circulation)

Update the circulation of the vortex blob.

Arguments

blob: The vortex blob.

new_circulation: The new circulation strength.

VEM.blob_circulation — Method

blob_circulation(blob::AbstractVortexBlob)

Arguments

blob: The vortex blob.

Returns

The circulation of the vortex blob.

VEM.blob_dimension — Method

blob_dimension(blob::AbstractVortexBlob)

Arguments

blob: The vortex blob.

Returns

The spatial dimension of the vortex blob.

VEM.blob_radius! — Method

blob_radius!(blob::AbstractVortexBlob, new_radius)

Update the core radius of the vortex blob.

Arguments

blob: The vortex blob.

new_radius: The new core radius.

VEM.blob_radius — Method

blob_radius(blob::AbstractVortexBlob)

Arguments

blob: The vortex blob.

Returns

The core radius of the vortex blob.

VEM.blob_scalar — Method

blob_scalar(blob::AbstractVortexBlob)

Arguments

blob: The vortex blob.

Returns

The floating-point type of the vortex blob data.

VEM.cells_per_axis — Method

cells_per_axis(mesh::CartesianMesh)

Get the number of cells per axis in the mesh.

Arguments

mesh: The Cartesian mesh.

Returns

The number of cells per axis in the Cartesian mesh.

VEM.current_time — Method

current_time(time::SimulationTime)

Compute the current simulation time based on the start_time, time_step, and current_step.

Arguments

time: An instance of SimulationTime.

Returns

The current simulation time as a Float.

VEM.diffusion! — Method

diffusion!(blobs, viscosity, mesh, time, time_step; time_scheme=ODE.RK4())

Apply diffusion to the given blobs over a time step using the specified time integration scheme.

Arguments

blobs: A collection of blobs.

viscosity: The fluid viscosity.

mesh: The Cartesian mesh whose nodes are the locations of the blobs.

time: The time at the start of the time step.

time_step: The time increment for diffusion.

time_scheme: (Optional) The time integration scheme to use, default is ODE.RK4().

VEM.direct_sum! — Method

direct_sum!(field, induced_field::AbstractInducedField, blobs, targets)

Compute the superposition of fields induced by a collection of blobs at a collection of targets (in-place).

Arguments

field: The superimposed induced field.

induced_field: The type of field to be induced by the blobs.

blobs: A single blob or an iterable collection of vortex blobs.

targets: A single target or an iterable collection of targets.

VEM.direct_sum — Method

direct_sum(induced_field::AbstractInducedField, blobs, targets)

Compute the superposition of fields induced by a collection of blobs at a collection of targets (out-of-place).

Arguments

induced_field: The type of field to be induced by the blobs.

blobs: A single blob or an iterable collection of vortex blobs.

targets: A single target or an iterable collection of targets.

Returns

A vector where each entry corresponds to the superposition of inductions at a target.

VEM.end_time — Method

end_time(time::SimulationTime)

Compute the end time of the simulation based on the start_time, time_step, and number_of_steps.

Arguments

time: An instance of SimulationTime.

Returns

The end time of the simulation as a Float.

VEM.evolve! — Method

evolve!(solver::LagrangianFlowSolver)

Evolve the flow field using the Lagrangian flow solver.

Arguments

solver: An instance of LagrangianFlowSolver containing the vortex blobs, flow properties, simulation time, and numerical configuration.

VEM.field_scalar — Method

field_scalar(::VelocityField, blob::AbstractVortexBlob)

Determine the scalar type of the velocity field induced by a vortex blob.

Arguments

::VelocityField: The type of field that is induced.

blob: The vortex blob for which the field is being induced.

Returns

The scalar type of the velocity field induced by the vortex blob.

VEM.induce — Method

induce(::VelocityField, blob::GaussianVortexBlob{2}, target)

Compute the velocity induced at target due to a 2D Gaussian vortex blob.

Arguments

VelocityField: The type of field that is induced.

blob: The vortex blob.

target: The target position.

Returns

The induced velocity at target.

VEM.induce — Method

induce(::VorticityField, blob::GaussianVortexBlob{2}, target)

Compute the vorticity induced at target due to a 2D Gaussian vortex blob.

Arguments

VorticityField: The type of field that is induced.

blob: The vortex blob.

target: The target position.

Returns

The induced vorticity at target.

VEM.interpolate_circulation — Method

interpolate_circulation(blobs, mesh::CartesianMesh, kernel::AbstractRedistributionKernel)

Interpolate the circulation of the blobs onto the nodes of the mesh using the redistribution kernel.

Arguments

blobs: The vortex blobs whose circulation is to be interpolated.

mesh: The Cartesian mesh onto which the circulation will be interpolated at the nodes.

kernel: The redistribution kernel that determines the weights for interpolation.

Returns

The interpolated circulation at the nodes of the Cartesian mesh.

VEM.mesh_bounds — Method

mesh_bounds(mesh::CartesianMesh)

Get the bounds of the mesh.

Arguments

mesh: The Cartesian mesh.

Returns

The bounds of the Cartesian mesh as a tuple of tuples, where each inner tuple contains the minimum and maximum bounds along each axis.

VEM.mesh_dimension — Method

mesh_dimension(::CartesianMesh)

Get the dimension of the CartesianMesh.

Arguments

::CartesianMesh: The Cartesian mesh.

Returns

The dimension of the Cartesian mesh.

VEM.mesh_nodes — Method

mesh_nodes(mesh)

Get the nodes of the mesh.

Arguments

mesh: The Cartesian mesh.

Returns

The nodes of the Cartesian mesh as a matrix of static vectors.

VEM.mesh_scalar — Method

mesh_scalar(::CartesianMesh)

Get the scalar type of the CartesianMesh.

Arguments

::CartesianMesh: The Cartesian mesh.

Returns

The scalar type of the Cartesian mesh.

VEM.node_spacing — Method

node_spacing(::CartesianMesh)

Get the spacing between nodes in the mesh.

Arguments

mesh: The Cartesian mesh.

Returns

The spacing between nodes in the Cartesian mesh.

VEM.nodes_per_axis — Method

nodes_per_axis(::CartesianMesh)

Get the number of nodes per axis in the mesh.

Arguments

mesh: The Cartesian mesh.

Returns

The number of nodes per axis in the Cartesian mesh.

VEM.population_control! — Method

population_control!(blobs, tolerance)

Prune vortex blobs based on their circulation magnitude to maintain a manageable population size. The pruning procedure ensures that the sum of circulation magnitude removed is smaller than the specified tolerance multiplied by the sum of all circulation magnitudes.

Arguments

blobs: A collection of vortex blobs.

tolerance: A scalar value that determines a relative threshold for pruning.

Returns

removed_circulation: The total circulation removed due to pruning.

VEM.redistribution — Method

redistribution(old_blobs, mesh::CartesianMesh, kernel::AbstractRedistributionKernel)

Redistribute the old_blobs onto the mesh using the redistribution kernel.

Arguments

old_blobs: The vortex blobs to be redistributed.

mesh: The Cartesian mesh onto which the blobs will be redistributed at the nodes.

kernel: The redistribution kernel that determines the weights for redistribution.

Returns

The redistributed vortex blobs at the nodes of the Cartesian mesh.

VEM.redistribution_weight — Method

redistribution_weight(::M4Prime, distance)

Compute the redistribution weight for a given distance using the M4' kernel.

Arguments

::M4Prime: An instance of the M4' redistribution kernel.

distance: The distance from the center of the vortex blob, normalized by the blob's core radius.

Returns

The redistribution weight corresponding to the given distance.

VEM.running — Method

running(time::SimulationTime)

Check if the simulation should continue running based on the current_step and number_of_steps.

Arguments

time: An instance of SimulationTime.

Returns

A boolean indicating whether the simulation should continue running (true) or has completed (false).

VEM.support_radius — Method

support_radius(::M4Prime)

Arguments

::M4Prime: An instance of the M4' redistribution kernel.

Returns

The support radius of the M4' redistribution kernel.