API

VEM.AbstractVortexBlob

VEM.GaussianVortexBlob

VEM.GaussianVortexBlob

VEM.GaussianVortexBlob

VEM.GaussianVortexBlob

VEM.blob_center

VEM.blob_center!

VEM.blob_circulation

VEM.blob_circulation!

VEM.blob_dimension

VEM.blob_radius

VEM.blob_radius!

VEM.blob_scalar

VEM.induced_velocity

VEM.induced_vorticity

VEM.induction_field_superposition

VEM.induction_field_superposition

VEM.induction_field_superposition

VEM.induction_field_superposition

VEM.induction_field_superposition!

VEM.induction_field_superposition!

VEM.induction_field_superposition!

VEM.induction_field_superposition!

VEM.AbstractVortexBlob — Type

AbstractVortexBlob{Dimension, Scalar <: AbstractFloat}

Abstract type representing a vortex blob in Dimension-dimensional space, parameterized by a floating-point type Scalar.

VEM.GaussianVortexBlob — Type

GaussianVortexBlob(circulation::Scalar, center::SA.SVector{Dimension, Scalar}, radius::Scalar) where {Dimension, Scalar}

A smooth vortex blob with a Gaussian distribution of vorticity. It is characterized by the amount of circulation it carries, a position and a finite core radius.

Arguments

circulation::Scalar: The circulation strength of the vortex blob.

center::SA.SVector{Dimension, Scalar}: The position of the blob.

radius::Scalar: The core radius of the blob.

Returns

A GaussianVortexBlob instance.

VEM.GaussianVortexBlob — Method

GaussianVortexBlob(circulation, center::AbstractVector, radius)

Constructor for GaussianVortexBlob when center is supplied as <:AbstractVector.

Arguments

circulation: The circulation strength of the vortex blob.

center::AbstractVector: The position of the blob.

radius: The core radius of the blob.

Returns

A GaussianVortexBlob instance.

VEM.GaussianVortexBlob — Method

GaussianVortexBlob(circulation, center::SA.SVector, radius)

Constructor for GaussianVortexBlob when center is supplied as SA.SVector.

Arguments

circulation: The circulation strength of the vortex blob.

center::SA.SVector: The position of the blob.

radius: The core radius of the blob.

Returns

A GaussianVortexBlob instance.

VEM.GaussianVortexBlob — Method

GaussianVortexBlob(circulation, center::Tuple, radius)

Constructor for GaussianVortexBlob when center is supplied as Tuple.

Arguments

circulation: The circulation strength of the vortex blob.

center::Tuple: The position of the blob.

radius: The core radius of the blob.

Returns

A GaussianVortexBlob instance.

VEM.blob_center! — Method

blob_center!(blob::AbstractVortexBlob, new_center)

Update the position of the vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

new_center: The new vortex blob position.

VEM.blob_center — Method

blob_center(blob::AbstractVortexBlob)

Return the position of the vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

Returns

The position of the vortex blob.

VEM.blob_circulation! — Method

blob_circulation!(blob::AbstractVortexBlob, new_circulation)

Update the circulation of the vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

new_circulation: The new circulation strength.

VEM.blob_circulation — Method

blob_circulation(blob::AbstractVortexBlob)

Return the circulation carried by the vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

Returns

The circulation of the vortex blob.

VEM.blob_dimension — Method

blob_dimension(blob::AbstractVortexBlob)

Return the spatial dimension of the given vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

Returns

The spatial dimension of the vortex blob.

VEM.blob_radius! — Method

blob_radius!(blob::AbstractVortexBlob, new_radius)

Update the core radius of the vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

new_radius: The new core radius.

VEM.blob_radius — Method

blob_radius(blob::AbstractVortexBlob)

Return the core radius of the vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

Returns

The core radius of the vortex blob.

VEM.blob_scalar — Method

blob_scalar(blob::AbstractVortexBlob)

Return the floating-point type used for the data of the given vortex blob.

Arguments

blob::AbstractVortexBlob: The vortex blob.

Returns

The floating-point type of the vortex blob data.

VEM.induced_velocity — Method

induced_velocity(blob::GaussianVortexBlob{2}, target)

Compute the velocity induced at target due to a 2D Gaussian vortex blob.

Arguments

blob::GaussianVortexBlob{2}: The vortex blob.

target: The target position.

Returns

The induced velocity at target.

VEM.induced_vorticity — Method

induced_vorticity(blob::GaussianVortexBlob{2}, target)

Compute the vorticity induced at target due to a 2D Gaussian vortex blob.

Arguments

blob::GaussianVortexBlob{2}: The vortex blob.

target: The target position.

Returns

The induced vorticity at target.

VEM.induction_field_superposition! — Method

induction_field_superposition!(field, induction::Function, blob::AbstractVortexBlob, target::AbstractVector{<:AbstractFloat})

Compute the value induced by a blob at a target in-place.

VEM.induction_field_superposition! — Method

induction_field_superposition!(field, induction::Function, blob::AbstractVortexBlob, targets)

Compute the field induced by a blob at targets in-place.

VEM.induction_field_superposition! — Method

induction_field_superposition!(field, induction::Function, blobs, target::AbstractVector{<:AbstractFloat})

Compute the superposition of fields induced by blobs at a single target in-place.

VEM.induction_field_superposition! — Method

induction_field_superposition!(field, induction::Function, blobs, targets)

Compute the superposition of fields induced by blobs at targets in-place.

Arguments

field: The superimposed induced field.

induction: A function that computes the value induced by a blob at a target.

blobs: An iterable collection of vortex blobs.

targets: An iterable collection of targets.

VEM.induction_field_superposition — Method

induction_field_superposition(induction::Function, blob::AbstractVortexBlob, target::AbstractVector{<:AbstractFloat})

Compute the value induced by a blob at a target.

VEM.induction_field_superposition — Method

induction_field_superposition(induction::Function, blob::AbstractVortexBlob, targets)

Compute the field induced by a blob at targets.

VEM.induction_field_superposition — Method

induction_field_superposition(induction::Function, blobs, target::AbstractVector{<:AbstractFloat})

Compute the superposition of fields induced by blobs at a single target.

VEM.induction_field_superposition — Method

induction_field_superposition(induction::Function, blobs, targets)

Compute the superposition of fields induced by blobs at targets.

Arguments

induction::Function: A function that computes the value induced by a blob at a target.

blobs: An iterable collection of vortex blobs.

targets: An iterable collection of targets.

Returns

A vector where each entry corresponds to the superposition of inductions at a target.